Make assistance amounts an integer

+++ db/warehouse/migrate/20200723144121_add_amount_to_direct_financial_assistance.rb
@@ -1,5 +1,5 @@
 class AddAmountToDirectFinancialAssistance < ActiveRecord::Migration[5.2]
   def change
-    add_column :direct_financial_assistances, :amount, :decimal
+    add_column :direct_financial_assistances, :amount, :integer
   end
 end
